Change: Use a `RaftConfig` trait to configure common types
Currently only the request and response type are configured, but other types can be now easily configured as well, such as `NodeId` (to be done in the next change) or `Runtime`.

Storage and network factory types are still provided as a separate type, since otherwise it would be hard to provide storage wrappers for testing.

This addresses https://github.com/datafuselabs/openraft/issues/204.